Fix warnings

+++ modules/beast_core/files/beast_FileOutputStream.h
@@ -53,7 +53,7 @@ public:
         @see TemporaryFile
     */
     FileOutputStream (const File& fileToWriteTo,
-                      int bufferSizeToUse = 16384);
+                      size_t bufferSizeToUse = 16384);
 
     /** Destructor. */
     ~FileOutputStream();
