+++ modules/beast_core/native/beast_posix_SharedCode.h
@@ -1327,7 +1327,7 @@ struct HighResolutionTimer::Pimpl : public Uncopyable
             shouldStop = false;
 
             if (pthread_create (&thread, nullptr, timerThread, this) == 0)
-                setThreadToRealtime (thread, newPeriod);
+                setThreadToRealtime (thread, (uint64) newPeriod);
             else
                 bassertfalse;
         }
